Added Password Encryption
Utilized the werkzeug library and the generate_password_has and check_password_hash functions to properly has the passwords to have ample amount of security. It's not a good idea to have plain-text passwords floating around in the database since it could potentially be exploited and attackers might access personal accounts.